+++ src/Services/Qrda/Cat1.php
@@ -20,11 +20,10 @@ use OpenEMR\Services\Qrda\Helpers\View;
 class Cat1 extends \Mustache_Engine
 {
     use Date;
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      use View;
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      use Cat1View;
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      use PatientView;
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      use Frequency;
-
+    use View;
+    use Cat1View;
+    use PatientView;
+    use Frequency;                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          use View;
 
     protected $templatePath =
         __DIR__ . DIRECTORY_SEPARATOR .
